feat(views): build the fullscreen player
