+++ crates/views/src/chrome/mod.rs
@@ -1,3 +1,4 @@
+mod aside;
 mod debug;
 mod player_bar;
 mod sidebar_left;
@@ -7,6 +8,7 @@ mod toasts;
 mod toolbar;
 pub(crate) mod tools;
 
+pub(crate) use aside::Aside;
 pub(crate) use debug::{FrameStats, Stats};
 pub(crate) use player_bar::PlayerBar;
 pub(crate) use sidebar_left::SidebarLeft;
