+++ crates/views/src/shared/mod.rs
@@ -11,6 +11,7 @@ pub(crate) mod picks;
 pub(crate) mod playlist_editor;
 pub(crate) mod shelves;
 pub(crate) mod tracks;
+pub(crate) mod transport;
 pub(crate) mod trouble;
 
 use gpui::prelude::*;
